+++ charts/gardener-extension-admission-gcp/charts/application/templates/rbac.yaml
@@ -9,10 +9,21 @@ rules:
 - apiGroups:
   - core.gardener.cloud
   resources:
+  - shoots
   - cloudprofiles
+  - secretbindings
   verbs:
   - get
   - list
+  - watch
+- apiGroups:
+  - ""
+  resources:
+  - secrets
+  verbs:
+  - get
+  - list
+  - watch
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
